moved email api functions to correct module path
